Correcion metodo velocidad y test

+++ src/model/SphereDistance.java
@@ -42,7 +42,7 @@ public class SphereDistance {
 		double x = (sphere1.getCoordinate().getxAxis()-sphere2.getCoordinate().getxAxis());
 		double y = (sphere1.getCoordinate().getyAxis()-sphere2.getCoordinate().getyAxis());
 		double z = (sphere1.getCoordinate().getzAxis()-sphere2.getCoordinate().getzAxis());
-		double aux = Math.sqrt(((Math.pow(x, 2))+(Math.pow(y, 2))+(Math.pow(z, 2))));
+		double aux = (int)Math.sqrt(((Math.pow(x, 2))+(Math.pow(y, 2))+(Math.pow(z, 2))));
 		averageDistance += aux ;
 		
 		return aux;
@@ -122,8 +122,8 @@ public class SphereDistance {
 
 //	public static void main(String[] args) {
 //		SphereDistance distance = new SphereDistance();
-//		Sphere a = new Sphere(new Coordinate(5,15,78.5),78);
-//		Sphere b = new Sphere(new Coordinate(10,30,63.4),78);
+//		Sphere a = new Sphere(new Coordinate(5,15,78),new Speed(10, 10, 10),15);
+//		Sphere b = new Sphere(new Coordinate(10,30,63),new Speed(10, 10, 10),15);
 //		System.out.println("Distancia :"+distance.calculateDistance(a,b));
 //	}
